+++ db/migrate/20131119160953_add_editor_to_site_users.rb
@@ -1,0 +1,5 @@
+class AddEditorToSiteUsers < ActiveRecord::Migration
+  def change
+    add_column :site_users, :editor, :boolean, null: false, default: false
+  end
+end