+++ test/controllers/project_users_controller_test.rb
@@ -6,17 +6,6 @@ class ProjectUsersControllerTest < ActionController::TestCase
     @project_user = project_users(:one)
   end
 
-  # test "should get index" do
-  #   get :index
-  #   assert_response :success
-  #   assert_not_nil assigns(:project_users)
-  # end
-  #
-  # test "should get new" do
-  #   get :new
-  #   assert_response :success
-  # end
-
   test "should resend project invitation" do
     post :resend, id: @project_user, format: 'js'
 
@@ -30,47 +19,6 @@ class ProjectUsersControllerTest < ActionController::TestCase
 
     assert_nil assigns(:project_user)
     assert_nil assigns(:project)
-    assert_response :success
-  end
-
-  test "should create project user" do
-    assert_difference('ProjectUser.count') do
-      post :create, project_user: { project_id: projects(:one).id, editor: true }, editors_text: users(:two).name + " [#{users(:two).email}]", format: 'js'
-    end
-
-    assert_not_nil assigns(:project_user)
-    assert_template 'index'
-  end
-
-  test "should create project user and automatically add associated user" do
-    assert_difference('ProjectUser.count') do
-      post :create, project_user: { project_id: projects(:single_design).id, editor: true }, editors_text: users(:pending).name + " [#{users(:pending).email}]", format: 'js'
-    end
-
-    assert_not_nil assigns(:project)
-    assert_not_nil assigns(:user)
-
-    assert_not_nil assigns(:project_user)
-    assert_template 'index'
-  end
-
-  test "should create project user invitation" do
-    assert_difference('ProjectUser.count') do
-      post :create, project_user: { project_id: projects(:one).id, editor: true }, editors_text: "invite@example.com", format: 'js'
-    end
-
-    assert_not_nil assigns(:project_user)
-    assert_not_nil assigns(:project_user).invite_token
-
-    assert_template 'index'
-  end
-
-  test "should not create project user with invalid project id" do
-    assert_difference('ProjectUser.count', 0) do
-      post :create, project_user: { project_id: -1, editor: true }, editors_text: users(:two).name + " [#{users(:two).email}]", format: 'js'
-    end
-
-    assert_nil assigns(:project_user)
     assert_response :success
   end
 
@@ -111,28 +59,13 @@ class ProjectUsersControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  # test "should show project_user" do
-  #   get :show, id: @project_user
-  #   assert_response :success
-  # end
-  #
-  # test "should get edit" do
-  #   get :edit, id: @project_user
-  #   assert_response :success
-  # end
-  #
-  # test "should update project_user" do
-  #   put :update, id: @project_user, project_user: @project_user.attributes
-  #   assert_redirected_to project_user_path(assigns(:project_user))
-  # end
-
   test "should destroy project user" do
     assert_difference('ProjectUser.count', -1) do
       delete :destroy, id: @project_user, format: 'js'
     end
 
     assert_not_nil assigns(:project)
-    assert_template 'index'
+    assert_template 'projects/members'
   end
 
   test "should allow viewer to remove self from project" do
@@ -141,7 +74,7 @@ class ProjectUsersControllerTest < ActionController::TestCase
     end
 
     assert_not_nil assigns(:project)
-    assert_template 'index'
+    assert_template 'projects/members'
   end
 
   test "should not destroy project user with invalid id" do